feat(file): 파일 존재 시 ConflictException 발생

+++ src/v1/apis/file/services/local-file.service.ts
@@ -1,6 +1,7 @@
 import FileService from './file.service.js';
 import path from 'path';
 import fs from 'fs';
+import { ConflictException } from 'src/v1/common/exceptions/core.error.js';
 
 export default class LocalFileService implements FileService {
   constructor(
@@ -18,6 +19,10 @@ export default class LocalFileService implements FileService {
   async upload(fileBuffer: Buffer, key: string): Promise<string> {
     const fullPath = path.join(this.baseDir, key);
     fs.mkdirSync(path.dirname(fullPath), { recursive: true });
+    if (fs.existsSync(fullPath)) {
+      throw new ConflictException('파일이 이미 존재합니다.');
+    }
+
     fs.writeFileSync(fullPath, fileBuffer);
     return this.getUrl(key);
   }
